# [Mono.Android] avoid System.Reflection.Emit usage for common calls