--- a/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Resources/Base/BuildReleaseArm64SimpleLegacy.apkdesc
+++ b/src/Xamarin.Android.Build.Tasks/Tests/Xamarin.ProjectTools/Resources/Base/BuildReleaseArm64SimpleLegacy.apkdesc
@@ -5,25 +5,25 @@
       "Size": 2604
     },
     "assemblies/Java.Interop.dll": {
-      "Size": 67759
+      "Size": 67924
     },
     "assemblies/Mono.Android.dll": {
-      "Size": 251399
+      "Size": 255765
     },
     "assemblies/mscorlib.dll": {
-      "Size": 769324
+      "Size": 769156
     },
     "assemblies/System.Core.dll": {
-      "Size": 28187
+      "Size": 28199
     },
     "assemblies/System.dll": {
-      "Size": 9171
+      "Size": 9180
     },
     "assemblies/UnnamedProject.dll": {
-      "Size": 2866
+      "Size": 2880
     },
     "classes.dex": {
-      "Size": 345208
+      "Size": 347796
     },
     "lib/arm64-v8a/libmono-btls-shared.so": {
       "Size": 1613872
@@ -32,16 +32,16 @@
       "Size": 707024
     },
     "lib/arm64-v8a/libmonodroid.so": {
-      "Size": 286736
+      "Size": 294736
     },
     "lib/arm64-v8a/libmonosgen-2.0.so": {
-      "Size": 4037584
+      "Size": 4030000
     },
     "lib/arm64-v8a/libxa-internal-api.so": {
       "Size": 65624
     },
     "lib/arm64-v8a/libxamarin-app.so": {
-      "Size": 20032
+      "Size": 20592
     },
     "META-INF/ANDROIDD.RSA": {
       "Size": 1213
@@ -74,5 +74,5 @@
       "Size": 1724
     }
   },
-  "PackageSize": 3983060
+  "PackageSize": 3991252
 }